requirements.txt
#pandas==2.2.3 # ajouter en local
mysqlclient==2.2.4
jupysql==0.10.14
#seaborn==0.13.2 # ajouter en local
grpcio==1.75.1
pymongo==4.10.1
pyspark==3.5.3
hdfs==2.7.3
minio==7.2.10
docker==7.1.0
kafka-python==2.0.2; python_version < '3.12'
kafka-python @ git+https://github.com/dpkp/kafka-python.git ; python_version >= '3.12'
python-dotenv==1.0.0
# Packages pour le notebook d'initialisation
pandas==2.1.4
seaborn==0.13.0
matplotlib==3.8.2
requests==2.31.0
delta-spark==3.3.2

# Dashboard packages
streamlit==1.31.0
deltalake==0.15.0
pyarrow==15.0.0
plotly==5.18.0